feat(home): vmux://home launcher with command-bar parity + default startup
New vmux://home Dioxus page: a centered Google-style launcher that reuses the
command bar full composition (live filter, file path completion + ghost text,
history suggestions, pages/tabs/spaces/commands) via a shared CommandPalette.

- Extract CommandPalette body into command_bar/palette.rs; the modal page keeps
  the backdrop shell, reveal handshake, sizing, and outside-pointer dismiss and
  renders CommandPalette; home renders it in a centered glass shell.
- Host: extract gather_command_bar_tabs + build_command_bar_open_payload from
  handle_open_command_bar; on_home_data_request replies with CommandBarOpenEvent
  to the requesting page; register "home" on the action/path/history emitters.
- Default startup_url -> vmux://home/ (empty + legacy vmux://agent/ retarget to home).
