+++ crates/vmux_desktop/src/lib.rs
@@ -132,6 +132,7 @@ impl Plugin for VmuxPlugin {
                 vmux_team::TeamPlugin,
                 vmux_history::HistoryPlugin,
                 vmux_agent::vibe::setup::AgentSetupPlugin,
+                vmux_layout::home::HomePlugin,
                 LayoutCefPlugin,
                 vmux_browser::ExtensionsPlugin,
                 BrowserPlugin,
