+++ crates/vmux_setting/src/settings.ron
@@ -1,6 +1,6 @@
 (
     browser: (
-        startup_url: "https://www.google.com",
+        startup_url: "vmux://home/",
     ),
     layout: (
         window: (
